added login logout

+++ app/view1/view1.js
@@ -14,11 +14,16 @@ angular.module('myApp.view1', ['ngRoute','core.keys'])
     controller: 'View1Ctrl'
 })
 
-.controller('View1Ctrl', ['$interval','$scope','Keys',function($interval,$scope,Keys) {
+.controller('View1Ctrl', ['$interval','$scope','Keys','$http','AuthenticationService','$rootScope',function($interval,$scope,Keys,$http,AuthenticationService, $rootScope) {
     var self = this;
     var totpObj = new TOTP();
 
-    getFromRemote(self, Keys, totpObj);
+    if (!window.localStorage.getItem('user'))
+        window.location.href="#!/login";
+
+    var authData = window.localStorage.getItem('authData');
+
+    getFromRemote(self, Keys, totpObj, authData);
 
 
 
@@ -29,22 +34,38 @@ angular.module('myApp.view1', ['ngRoute','core.keys'])
 
 
     self.delete = function(){
+
+        var selectedItems = self.keys.filter(function(elem,index,self){return elem.selected});
+        if (!confirm("Delete " + selectedItems.length + " items?" )){
+            return;
+        }
+
+        self.keys = self.keys.filter(function(elem,index,self){
+            return !selectedItems.includes(elem);
+        });
+        angular.forEach(selectedItems,function(item){
+            Keys.restricted().delete({id: item.id}).$promise.then(function(response){
+                console.log(response);
+            });
+        });
+        /*
         var notDeletedItems = self.keys.filter(function(elem,index,self){return !elem.selected;});
         window.localStorage.setItem("keys",JSON.stringify(notDeletedItems));
         self.keys = notDeletedItems;
         if(self.keys.length < 1)
             window.location.href="#!/view2";
-    
+            */
+
     }
 
     self.checked = function(item){
-    
+
         self.selected = self.keys.filter(function(elem,index,self){return elem.selected}).length>0;
     }
 
     self.copy = function(item){
-        
-        
+
+
         var lblInfo;
         if( copyToClipboard(item.totp)) {
             lblInfo = "Copied.";
@@ -59,11 +80,11 @@ angular.module('myApp.view1', ['ngRoute','core.keys'])
             if (self.secs == 0){
                 self.showInfoLabel = false;
                 $interval.cancel(labelTimer);
-            
+
             }
-        
+
         },1000);
-    
+
     };
 
 
@@ -100,6 +121,10 @@ angular.module('myApp.view1', ['ngRoute','core.keys'])
     };
 
 
+    self.logout = function(){
+        AuthenticationService.logout();
+        window.location.href="#!/login";
+    }
 
 
 
@@ -110,10 +135,10 @@ var updateKeys = function(keys){
         angular.forEach(keys,function(result){
             result.time = totpObj.getTime();
             result.totp = totpObj.getOTP(result.secret);
-            if (result.time == 30){
+            if (result.time === 30){
                 result.totp = totpObj.getOTP(result.secret);
             }
-        
+
         });
 
 }
@@ -129,8 +154,9 @@ var addItem = function(obj){
 
 };
 
-var getFromRemote = function(self, Keys, totpObj) {
-    Keys.query().$promise.then(function(results){
+var getFromRemote = function(self, Keys, totpObj,authData) {
+    Keys.restricted(authData).query().$promise.then(function(results){
+        console.log(results);
         if (results.length < 1){
             window.location.href="#!/view2";
         } else {
@@ -142,6 +168,10 @@ var getFromRemote = function(self, Keys, totpObj) {
             self.keys= results;
 
         }
+    }, function error(response){
+        console.log(response.statusText);
+        self.logout();
+        window.location.href="#!/login";
     });
 
 };
